+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added support for Google analytics
 - Improved API security
 - Implemented a unique auto generate 32 bit key for API access
+- Fixed synonym search
 
 ### Version 12.5
 - Code clean-up